Fix pixmap compilation with clang
Clang does not allow a variable to be used for the "K" register, even if
all uses of the inlined function use a constant value.

The current version of pixmap doesn't use the inline function. It uses a
macro regardless of the optimization level. There should be no
performance overhead from using the macro.